+++ src/main/scala/swiss/dasch/domain/MovingImageService.scala
@@ -51,8 +51,8 @@ case class MovingImageService(storage: StorageService, executor: CommandExecutor
       _                          <- ZIO.logInfo(s"Extracting metadata for $derivative, $assetRef")
       ffprobeInfo                <- extractWithFfprobe(derivative)
       (dimensions, duration, fps) = ffprobeInfo
-      originalMimeType           <- mimeTypeGuesser.guess(Path(original.originalFilename.value))
       derivativeMimeType         <- mimeTypeGuesser.guess(derivative.toPath)
+      originalMimeType           <- mimeTypeGuesser.guess(Path(original.originalFilename.value))
     } yield MovingImageMetadata(dimensions, duration, fps, derivativeMimeType, originalMimeType)
 
   private def extractWithFfprobe(derivative: MovingImageDerivativeFile) = for {
